Add in-kernel init.d support
Systemless init.d (except userspace to store scripts :P)

The script will check for init.d path inside /system/etc, if it's not found tries to create the folder.
If it's still not available means that can't write to /system even if we mounted it as r/w in fstab, therefore write in /data/local instead.

System mounted as r/w is mandatory to be written.